TST: Use DATALAD_HIRNI_TOOLBOX_URL with demo dataset
That config already is respected by the cfg_hirni procedure.
However, the demo dataset we provide on github has the
toolbox already registered, which is why we need to rewrite
its URL when installing it, in order for the tests to truly
use a different (local) to clone the toolbox from.