TST: Use cached datasets
Get a copy of cached dataset decorators from datalad core and fix them.
That issue being: AnnexRepo.get() doesn't take a list when key=True.
Working around here in the dataset caching helpers, so datalad-hirni and
issue 158 don't have to wait for yet another release of datalad.
Ultimately this needs fixing in core and as soon as there's a release w/
that fix, we should import them from core and delete it hirni.

This should address issue 158, since it's happening only, if we keep
flooding the original source with requests. Although it's not entirely
clear how exactly that is unfolding on the remote end, it's clear that
this "fixes" the issue here (or rather in datalad-extensions) and
furthermore this "solution" makes a whole lot of sense in terms of
runtime and traffic for the tests regardless of the actual issue.

(Closes #158)